Added My Output Images from the PyTorch-YOLOv3 Model
Added output imaged from PyTorch-YOLOv3 model as well as started a ProjectPart3Notebook for our deliverable. Add the information about your models into there when you can. If you are able to collect/determine some metrics for comparison or Vinith with try to use the images to get the accuray of the model/precision/recall or those types of metrics. But my model doesn't spit those out like Matt Thayer's.